f2fstat: fix command-line options
In ARM, the char type is always defined as unsigned (which is allowed
by the C standard).  This caused the terminator return value of getopt (-1)
was being converted to 255 and thus never matched the loop exit condition.

Also printfs warning 'unknown option' has been removed from all the tools
because getopt() itself warns with a message.